move create_roi_plot to plotting module
