+++ src/ophys_etl/modules/segmentation/graph_utils/feature_vector_segmentation.py
@@ -1,5 +1,3 @@
-import matplotlib.figure as mplt_figure
-
 from typing import Optional, List, Tuple
 import networkx as nx
 import numpy as np
@@ -15,6 +13,9 @@ from ophys_etl.modules.decrosstalk.ophys_plane import OphysROI
 
 from ophys_etl.modules.segmentation.graph_utils.feature_vector_rois import (
     PearsonFeatureROI)
+
+from ophys_etl.modules.segmentation.graph_utils.plotting import (
+    create_roi_plot)
 
 import logging
 
@@ -303,59 +304,6 @@ def convert_to_lims_roi(origin: Tuple[int, int],
                      valid=False,
                      mask=[i.tolist() for i in new_mask])
     return roi
-
-
-def create_roi_plot(plot_path: pathlib.Path,
-                    img_data: np.ndarray,
-                    roi_list: List[ExtractROI]) -> None:
-    """
-    Generate a side-by-side plot comparing the image data
-    used to seed ROI generation with the borders of the
-    discovered ROIs
-
-    Parameters
-    ----------
-    plot_path: pathlib.Path
-        Path to file where plot will be saved
-
-    img_data: np.ndarray
-        The baseline image over which to plot the ROIs
-
-    roi_list: List[ExtractROI]
-
-    Returns
-    -------
-    None
-    """
-    fig = mplt_figure.Figure(figsize=(40, 20))
-    axes = [fig.add_subplot(1, 2, i) for i in [1, 2]]
-    axes[0].imshow(img_data)
-    axes[1].imshow(img_data)
-
-    bdry_pixels = np.zeros(img_data.shape, dtype=int)
-    for roi in roi_list:
-        ophys_roi = OphysROI(
-                        roi_id=0,
-                        x0=roi['x'],
-                        y0=roi['y'],
-                        width=roi['width'],
-                        height=roi['height'],
-                        valid_roi=False,
-                        mask_matrix=roi['mask'])
-
-        bdry = ophys_roi.boundary_mask
-        for ir in range(ophys_roi.height):
-            for ic in range(ophys_roi.width):
-                if bdry[ir, ic]:
-                    bdry_pixels[ir+ophys_roi.y0,
-                                ic+ophys_roi.x0] = 1
-
-    bdry_pixels = np.ma.masked_where(bdry_pixels == 0,
-                                     bdry_pixels)
-    axes[1].imshow(bdry_pixels, cmap='autumn', alpha=1.0)
-    fig.tight_layout()
-    fig.savefig(plot_path)
-    return None
 
 
 class FeatureVectorSegmenter(object):
